+++ lib/workato/connector/sdk/connection.rb
@@ -30,10 +30,13 @@ module Workato
         RefreshOutput = T.type_alias do
           T.any(
             [
-              ActiveSupport::HashWithIndifferentAccess,
-              T.nilable(String)
+              ActiveSupport::HashWithIndifferentAccess, # tokens
+              T.nilable(ActiveSupport::HashWithIndifferentAccess) # settings
             ],
-            ActiveSupport::HashWithIndifferentAccess
+            [
+              ActiveSupport::HashWithIndifferentAccess # tokens
+            ],
+            ActiveSupport::HashWithIndifferentAccess # tokens
           )
         end
       end