+++ lib/workato/extension/string.rb
@@ -155,22 +155,30 @@ module Workato
       end
 
       def hmac_sha256(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('sha256')
         Types::Binary.new(::OpenSSL::HMAC.digest(digest, key, self))
       end
 
       def hmac_sha512(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('sha512')
         Types::Binary.new(::OpenSSL::HMAC.digest(digest, key, self))
       end
 
       def rsa_sha256(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('sha256')
         private_key = ::OpenSSL::PKey::RSA.new(key)
         Types::Binary.new(private_key.sign(digest, self))
       end
 
       def rsa_sha512(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('sha512')
         private_key = ::OpenSSL::PKey::RSA.new(key)
         Types::Binary.new(private_key.sign(digest, self))
@@ -187,17 +195,32 @@ module Workato
       end
 
       def hmac_sha1(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('sha1')
         Types::Binary.new(::OpenSSL::HMAC.digest(digest, key, self))
       end
 
       def hmac_md5(key)
+        assert_string_argument!(key, 'key')
+
         digest = ::OpenSSL::Digest.new('md5')
         Types::Binary.new(::OpenSSL::HMAC.digest(digest, key, self))
       end
 
       def from_xml
         Workato::Utilities::Xml.parse_xml_to_hash(self)
+      end
+
+      private
+
+      def assert_string_argument!(value, arg_name)
+        return if value.is_a?(String)
+
+        Kernel.raise(
+          Workato::Connector::Sdk::ArgumentError,
+          "Expected a String for '#{arg_name}' parameter, given #{value.class.name}"
+        )
       end
     end
   end